feat: Add piscine support in calculator
